# docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Tester.com API
  version: 1.0.0
  description: |
    O amazonQA é uma Aplicação completa com API REST gratuita feita com next.js, que simula uma loja virtual com intuito de servir de material de estudos de testes de API e Web.
    Não deixe de seguir o autor do projeto Reinaldo M R Junior no LinkedIn: https://www.linkedin.com/in/reinaldo-mateus-rossetti-a28a5423/
    E deixar um star no repositório: https://github.com/reinaldorossetti/amazonQA.com
servers:
  - url: http://127.0.0.1:3001
    description: Backend local (Next.js)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
          format: float
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        manufacturer:
          type: string
          nullable: true
        line:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
...
security: []
paths:
  /api/products:
    get:
      tags: [Products]
      summary: "Listar produtos (Acesso: público)"
      parameters:
        - in: query
          name: category
          schema:
            type: string
      responses:
        '200':
          description: Lista de produtos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Products]
      summary: "Criar produto (Acesso: admin)"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Produto criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Dados inválidos
        '401':
          description: Não autenticado
        '403':
          description: Requer admin
  /api/products/{id}:
    get:
      tags: [Products]
      summary: "Buscar produto por ID (Acesso: público)"
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Produto não encontrado

    put:
      tags: [Products]
      summary: "Atualizar produto (Acesso: admin)"
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Produto atualizado
        '401':
          description: Não autenticado
        '403':
          description: Requer admin
        '404':
          description: Produto não encontrado

    delete:
      tags: [Products]
      summary: "Excluir produto (Acesso: admin)"
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Produto removido
        '401':
          description: Não autenticado
        '403':
          description: Requer admin
        '404':
          description: Produto não encontrado
...